requirements.txt
RPi.GPIO==0.7.0
numpy==1.21.0
scipy==1.7.0